requirements.txt
appdirs==1.4.3
asgi-redis==1.3.0
asgiref==1.1.2
attrs==17.1.0
autobahn==17.5.1
Automat==0.6.0
automium==0.2.6
automium-web==0.1.1
Click==7.0
constantly==15.1.0
Dajax==1.3
deap==1.2.2
Django==2.1.5
django-crispy-forms==1.7.2
django-dajax==0.9.2
django-dajaxice==0.7
django-email==0.1.10
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-widget-tweaks==1.4.3
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
legacy==0.1.6
markdown2==2.3.7
MarkupSafe==1.1.0
msgpack-python==0.5.6
packaging==18.0
Pillow==5.4.0
PyMySQL==0.9.3
pyparsing==2.3.1
pytz==2017.2
redis==3.0.1
reportlab==3.4.0
six==1.12.0
txaio==18.8.1
virtualenv==16.2.0
Werkzeug==0.14.1
